# Optimize logical/visual children iteration on the hot path