## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
Removes boxing allocations from several fairly hot paths. Memory traffic results below are from `ControlCatalog` startup.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
![image](https://user-images.githubusercontent.com/2588062/99186416-6331b100-2750-11eb-86b5-79615dfbdbda.png)
![image](https://user-images.githubusercontent.com/2588062/99186432-8eb49b80-2750-11eb-98cb-757cda374495.png)


## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
![image](https://user-images.githubusercontent.com/2588062/99186417-67f66500-2750-11eb-8b40-9adce1b63450.png)
![image](https://user-images.githubusercontent.com/2588062/99186435-93794f80-2750-11eb-9669-44b13651a632.png)


## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->

For people wondering if a single boxing allocation for an enumerator matters - raw throughput comparison for going over an `AvaloniaList<int>` (just 4 elements).

|        Method |     Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |---------:|---------:|---------:|------:|-------:|------:|------:|----------:|
|    WithBoxing | 69.82 ns | 1.399 ns | 2.414 ns |  1.00 | 0.0095 |     - |     - |      40 B |
| WithIteration | 12.89 ns | 0.268 ns | 0.251 ns |  0.19 |      - |     - |     - |         - |


## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Avaloniaui.net with user documentation

## Breaking changes
<!--- List any breaking changes here. When the PR is merged please add an entry to https://github.com/AvaloniaUI/Avalonia/wiki/Breaking-Changes -->


## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->